fix(core): turn all dependencies to be peer, target the issue #142
Currently the README declares that the consumer should install the dependencies manually, however the same (but not the same actually, see below) libs are claimed to be dependencies.

In fact, both dependencies are being installed. Webpack ignores the dependencies from this package.json anyway, however npm still shows it in audits etc.

The correct approach would be to let users decide on particular versions of the libraries this package depends on => all should be peer.